refactor(daemon): extract run-event analytics scanners into run-event-analytics.ts
Strangler-fig slice 3: move the pure run-event analytics/retry scanners out
of server.ts (god file) into a new sibling module, byte-identical bodies.

Moved: resolveRunProjectKindForAnalytics, scanRunEventsForFinishedProps,
scanRunEventsForRetrySideEffects, fileNameFromToolInputPath,
filesystemWriteFileNamesFromRunEvents, filesystemEmptyAnswerFallbackText,
retryFinalResultForRunStatus, runRetryEventsForAnalytics, plus their seven
__forTest* wrappers. All are pure (read only their args + imported helpers,
no daemon module state).

server.ts imports back the five functions it references and re-exports the
seven __forTest* wrappers so its public surface (the daemon test suite imports
them from server.ts) is unchanged. server.ts -147 net lines (5603 -> 5434).

Validation: tsc (src+tests) clean; pnpm guard 55/0; chat-run-sse-shapes char
suite + run-lifecycle-analytics + tool-loop-persistence 29/29 green; orphan-ref
and duplicate-binding checks clean (server.ts is @ts-nocheck so tsc can't see
moved-symbol breakage).
